# Optimize Vec allocations in map deserialization.