Before:

utils::bench_deserialize_map       ... bench:   171.39 ns/iter (+/- 3.53)
utils::bench_deserialize_map_bytes ... bench:   148.27 ns/iter (+/- 3.19)

After:

utils::bench_deserialize_map       ... bench:   168.34 ns/iter (+/- 3.19)
utils::bench_deserialize_map_bytes ... bench:   145.40 ns/iter (+/- 1.92)